# Conflicts:
#	.idea/workspace.xml
#	src/views/landing.html

+++ src/assets/js/controllers/dashboardController.js
@@ -8,6 +8,7 @@ class DashboardController {
     constructor() {
         this.gameRepository = new GameRepository();
         this.userRepository = new UserRepository();
+        this.generalRepository = new GeneralRepository();
 
         $.get("views/dashboard.html")
          .done((data) => this.setup(data))
@@ -16,9 +17,26 @@ class DashboardController {
 
     async setup(data) {
         this.dashboardView = $(data);
+        this.user          = await this.userRepository.get(sessionManager.get("userID"));
         this.games         = await this.gameRepository.getAll();
         this.materials     = await this.gameRepository.getMaterials();
         this.grades        = await this.gameRepository.getGrades();
+
+        // statistics
+        const averageRating = await this.generalRepository.getAverageRating();
+        const amountUsers = await this.generalRepository.getAmountUsers();
+        const amountFavorites = await this.generalRepository.getAmountFavorites();
+
+        console.log(averageRating[0].averageRating);
+        console.log(amountUsers[0].amountUsers);
+        console.log(amountFavorites[0].amountFavorites);
+
+        this.dashboardView.find("#adminHeading").html("Welkom " + this.user.firstname);
+
+        this.dashboardView.find("#averageRating").html(averageRating[0].averageRating.toFixed(2));
+        this.dashboardView.find("#amountUsers").html(amountUsers[0].amountUsers);
+        this.dashboardView.find("#amountFavorites").html(amountFavorites[0].amountFavorites);
+
 
         this.handleRenderMasonry(this.games);
         this.dashboardView.find('#uploadMainImg').on('change', (e) => {
@@ -41,16 +59,22 @@ class DashboardController {
         this.dashboardView.find('#updateGame').on('click', (e) => {
             this.handleUpdateGameClick(e)
         });
+
         this.dashboardView.find('#update-form').submit((e) => {
             this.updateGame(e)
         });
 
-        this.dashboardView.find('#create-form').submit((e) => {
+        this.dashboardView.find('.game-add').on("click", (e) => {
             this.createGame(e)
         });
 
+        this.dashboardView.find(".addUpdateRules").on("click", (e) => {
+            this.handleAddRule(e);
+        });
 
-
+        this.dashboardView.find(".resetRules").on("click", (e) => {
+            this.handleResetRules(e)
+        });
 
 
         $(".content").empty().append(this.dashboardView);
@@ -120,22 +144,80 @@ class DashboardController {
         try {
 
             const gameRepository = new GameRepository();
-            const deleteGame     = await gameRepository.delete($(this).attr('data-id'));
 
-            new DashboardController()
+            notificationManager.confirm('Weet je zeker dat u dit wilt verwijderen?', async () => {
+                await gameRepository.delete($(this).attr('data-id'));
+                new DashboardController()
+            });
+
         } catch (e) {
             console.log(e);
             notificationManager.alert("error", 'Er is wat misgegaan...');
         }
     }
 
+    /**
+     * Handles adding a rule
+     * @param e
+     * @returns {Promise<void>}
+     */
+    async handleAddRule(e) {
+        try {
+            const gameRepository = new GameRepository();
 
+            let value = notificationManager.input("Maak een regel aan!", async (input) => {
+                console.log($(e.target).attr('data-id'));
+                console.log(input);
+
+                if (input) {
+                    let body = {
+                        gameID:      $(e.target).attr('data-id'),
+                        description: input
+                    };
+
+                    await gameRepository.addRules(body);
+
+                    const gameRules = await gameRepository.getRules($(e.target).attr('data-id'));
+
+                    this.dashboardView.find(".gameUpdateRules").html("");
+                    this.dashboardView.find(".gameUpdateRules").append(gameRules.map(RuleTableItem));
+
+                    notificationManager.alert("success", "Nieuwe regel is toegevoegd!")
+                } else {
+                    notificationManager.alert("error", 'Vul een regel in.');
+                }
+            });
+        } catch (e) {
+            console.log(e);
+            notificationManager.alert("error", 'Er is wat misgegaan...');
+        }
+    }
+
+    async handleResetRules(e) {
+        try {
+
+            await this.gameRepository.deleteRules({
+                gameID: $(e.target).attr('data-id')
+            });
+
+            await this.gameRepository.getRules($(e.target).attr('data-id'));
+            this.dashboardView.find(".gameUpdateRules").html("");
+            notificationManager.alert("success", "Regels zijn gereset!")
+
+
+        } catch (e) {
+            console.log(e);
+            notificationManager.alert("error", 'Er is wat misgegaan...');
+        }
+    }
 
     async handleUpdateGameClick(e) {
         try {
             const gameRepository = new GameRepository();
             const game           = await gameRepository.get($(e.target).attr('data-id'));
             const gameMaterial   = await gameRepository.getMaterial($(e.target).attr('data-id'));
+            const gameRules      = await gameRepository.getRules($(e.target).attr('data-id'));
+
 
             this.dashboardView.find("[name='update-id']").val(game.gameID);
             this.dashboardView.find("[name='update-name']").val(game.title);
@@ -144,6 +226,11 @@ class DashboardController {
             this.dashboardView.find("[name='update-type']").val(game.type);
             this.dashboardView.find("[name='update-grade").val(game.gradeID);
 
+            this.dashboardView.find(".gameUpdateRules").html("");
+            this.dashboardView.find(".gameUpdateRules").append(gameRules.map(RuleTableItem));
+            this.dashboardView.find(".addUpdateRules").attr("data-id", $(e.target).attr('data-id'));
+            this.dashboardView.find(".resetRules").attr("data-id", $(e.target).attr('data-id'));
+
             for (var index = 0; index < gameMaterial.length; index++) {
                 this.dashboardView.find("option[value='" + gameMaterial[index].materialID + "'").attr("selected", "selected");
             }
@@ -151,8 +238,8 @@ class DashboardController {
 
         } catch (e) {
             console.log(e);
+            notificationManager.alert("error", 'Er is wat misgegaan...');
         }
-
     }
 
     async updateGame(e) {
@@ -172,10 +259,9 @@ class DashboardController {
             console.log(body);
             await this.gameRepository.update(body);
             notificationManager.alert("success", this.dashboardView.find("[name='update-name']").val() + " is bijgewerkt!");
-        } catch(e) {
+        } catch (e) {
             console.log(e);
             notificationManager.alert("error", 'Er is wat misgegaan...');
-
         }
     }
 
@@ -183,22 +269,34 @@ class DashboardController {
         try {
             e.preventDefault();
 
+            let imageName = this.dashboardView.find("[name='create-name']").val().replace(/\s+/g, '-').toLowerCase();
+
             let body = {
-                title:       this.dashboardView.find("[name='create-name']").val(),
-                description: this.dashboardView.find("[name='create-description']").val(),
-                imageUrl: "./assets/img/games/placeholder.png",
-                floorplanUrl: "./assets/img/games/floorplan/placeholder.png",
-                minPlayers:  this.dashboardView.find("[name='create-minPlayers']").val(),
-                type:        this.dashboardView.find("[name='create-type']").val(),
-                gradeID:     this.dashboardView.find("[name='create-grade").val(),
-                materials:   this.dashboardView.find("[name='create-material").val()
+                title:        this.dashboardView.find("[name='create-name']").val(),
+                description:  this.dashboardView.find("[name='create-description']").val(),
+                imageUrl:     "./assets/img/games/" + imageName + ".jpg",
+                floorplanUrl: "./assets/img/games/floorplan/" + imageName + ".jpg",
+                minPlayers:   this.dashboardView.find("[name='create-minPlayers']").val(),
+                type:         this.dashboardView.find("[name='create-type']").val(),
+                gradeID:      this.dashboardView.find("[name='create-grade").val(),
+                materials:    this.dashboardView.find("[name='create-material").val(),
             };
 
-            console.log(body);
-            await this.gameRepository.create(body);
-            notificationManager.alert("success",  this.dashboardView.find("[name='create-name']").val() + " is toegevoegd!");
+            this.dashboardView.find("#mainImgUrl").val("assets/img/games/" + imageName + ".jpg");
+            this.dashboardView.find("#floorplanImgUrl").val("assets/img/games/floorplan/" + imageName + ".jpg");
 
-        } catch(e) {
+            const newGame = await this.gameRepository.create(body);
+
+            this.dashboardView.find("#create-form").submit();
+
+
+            notificationManager.alert("success", this.dashboardView.find("[name='create-name']").val() + " is toegevoegd!");
+
+            notificationManager.confirm('Ga naar nieuwe spel', async () => {
+                new GameDetailController(newGame.insertId)
+            });
+
+        } catch (e) {
             console.log(e);
             notificationManager.alert("error", 'Er is wat misgegaan...');
